C&P'd
Dr. Hills, 

I just C&P'd into this repository from the other one on my account. I do not know why it made a second repository when I pushed my changes. I can't access that one if I'm not logged in but it's not a private repository so I do not know what happened.

-Jared